refactor(src): Write a function to reuse preconditions in gsl_Expects
Reference: #1

+++ src/viewport.cpp
@@ -1,5 +1,7 @@
 #include "viewport.hpp"
 
+#include <cstddef>
+
 namespace raytracing {
 Viewport::Viewport(
 	const Eigen::Matrix3f &mat) // NOLINT(modernize-pass-by-value)
@@ -10,7 +12,7 @@ Viewport::Viewport(Vector3fConstRef delta_u, Vector3fConstRef delta_v,
 	mat_ << delta_u, delta_v, pixel00_loc;
 }
 
-Point3f Viewport::at(float i, float j) const {
+Point3f Viewport::at(size_t i, size_t j) const {
 	return mat_ * Vector3f{i, j, 1};
 }
